Grond-main 2/requirements.txt
pytz
pandas
numpy
requests
flask
gunicorn
python-telegram-bot>=20.0
polygon-api-client
websocket-client==1.6.0
prometheus_client
pandas_market_calendars
scikit-learn==1.7.1
cvxpy
PyPortfolioOpt
riskfolio-lib
backtrader
zipline-reloaded
ib_insync
jax[cpu]
QuantLib-Python
pytest
flake8
python-dotenv
optuna
gymnasium 
stable-baselines3>=2.0.0
xgboost==2.0.3
uvicorn
fastapi
joblib==1.3.2
boto3
